+++ demo/hadamard_adder.py
@@ -1,3 +1,4 @@
+import numpy as np
 from braket.devices import LocalSimulator
 from braket.ir.openqasm import Program
 
@@ -9,10 +10,6 @@ adder_qasm = """
  */
 OPENQASM 3;
 include "stdgates.inc";
-
-output bit[4] a_in;
-output bit[4] b_in;
-output bit[5] ans;
 
 gate majority a, b, c {
     cx c, b;
@@ -28,16 +25,9 @@ gate unmaj a, b, c {
 
 qubit cin;
 qubit[4] a;
-qubit[4] b;
 qubit[4] b_copy;
 qubit cout;
-
-// initialize qubits
-reset cin;
-reset a;
-reset b;
-reset b_copy;
-reset cout;
+qubit[4] b;
 
 // set input states
 for i in [0: 3] {
@@ -52,23 +42,23 @@ for i in [3: -1: 1] { majority a[i], b[i - 1], a[i - 1]; }
 cx a[0], cout;
 for i in [1: 3] { unmaj a[i], b[i - 1], a[i - 1]; }
 unmaj cin, b[3], a[3];
-
-// measure results
-a_in = measure a;
-b_in = measure b_copy;
-ans[0] = measure cout;
-ans[1:4] = measure b[0:3];
 """
 
 
-device = LocalSimulator("braket_oq3_native_sv")
+device = LocalSimulator("braket_oq3_sv")
 
 adder = Program(source=adder_qasm)
-num_shots = 10
+num_shots = 10000
 result = device.run(adder, shots=num_shots).result()
 
+
+def bit_string_to_int(bit_string):
+    return bit_string @ (2 ** np.arange(bit_string.size)[::-1])
+
+
 for shot in range(num_shots):
-    a = result.output_variables["a_in"][shot]
-    b = result.output_variables["b_in"][shot]
-    ans = result.output_variables["ans"][shot]
-    print(f"{int(a, base=2)} + {int(b, base=2)} = {int(ans, base=2)}")
+    measurement = result.measurements[shot]
+    a = measurement[1:5]
+    b = measurement[5:9]
+    ans = measurement[9:]
+    print(f"{bit_string_to_int(a)} + {bit_string_to_int(b)} = {bit_string_to_int(ans)}")
